update to reflect change to ScannableStatus
PVScananbleBean in this repository is the only place in the whole of GDA
where the getScannableName() method of ScannableStatus is used. It's use
here is redundant, and because of this ScannableStatus (in gda-core) can
be simplified to a simple enum.

The use of ScannableStatus.getScannableName() is redundant in
PVScannableBean as it is only used to check that the name of the
scannable in the ScannableStatus was the one that this bean was added as
a listener to in the startDisplay() method below. This should always be
the case.

Additionally, this is legacy code that should no longer be used.

Change-Id: Ide904f616646e095ecea1a745800cb7a160b7703